+++ packages/vendure-plugin-admin-social-auth/src/admin-social-auth.plugin.ts
@@ -34,7 +34,8 @@ export interface AdminSocialAuthPluginOptions {
     }
     return config;
   },
-  compatibility: '>=2.2.0',
+  dashboard: './dashboard/index.tsx',
+  compatibility: '>=3.2.0',
 })
 export class AdminSocialAuthPlugin {
   static options: AdminSocialAuthPluginOptions;